Move EpochSchedule into own module
